requirements.txt
Pillow
rectangle-packer